реализована генерация файлов проектов для socket.io.client-cpp

+++ libCliqueWebRTC.SetUp/main.py
@@ -24,18 +24,24 @@ config = {
     , "source_tag"   : "boost-1.64.0"                      # the tag for the branch to create from
     , "build_targets": # a set of build configurations
         [#targets
-            { "args":"--with-system --with-date_time --with-random --with-regex link=static runtime-link=static threading=multi address-model=32" # how to build
-            , "src": "stage/lib"                                                                                                                  # libraries are here
-            , "dst": "stage/lib_Win32"}                                                                                                           # move them here
-        ,   { "args":"--with-system --with-date_time --with-random --with-regex link=static runtime-link=shared threading=multi address-model=32" #
-            , "src": "stage/lib"                                                                                                                  #
-            , "dst": "stage/lib_Win32"}                                                                                                           #
-        ,   { "args":"--with-system --with-date_time --with-random --with-regex link=static runtime-link=static threading=multi address-model=64" #
-            , "src": "stage/lib"                                                                                                                  #
-            , "dst": "stage/lib_Win64"}                                                                                                           #
-        ,   { "args":"--with-system --with-date_time --with-random --with-regex link=static runtime-link=shared threading=multi address-model=64" #
-            , "src": "stage/lib"                                                                                                                  #
-            , "dst": "stage/lib_Win64"}                                                                                                           #
+            { "command" : "b2.exe"                                                                                                                     # this step is absolutely mandatory for modular boost
+            , "args"    : "headers" }                                                                                                                  # i.e. taken from git directly
+        ,   { "command" : "b2.exe"
+            , "args"    : "--with-system --with-date_time --with-random --with-regex link=static runtime-link=static threading=multi address-model=32" # how to build
+            , "src"     : "stage/lib"                                                                                                                  # libraries are here
+            , "dst"     : "stage/lib_Win32" }                                                                                                          # move them here
+        ,   { "command" : "b2.exe"
+            , "args"    : "--with-system --with-date_time --with-random --with-regex link=static runtime-link=shared threading=multi address-model=32" #
+            , "src"     : "stage/lib"                                                                                                                  #
+            , "dst"     : "stage/lib_Win32" }                                                                                                          #
+        ,   { "command" : "b2.exe"
+            , "args"    : "--with-system --with-date_time --with-random --with-regex link=static runtime-link=static threading=multi address-model=64" #
+            , "src"     : "stage/lib"                                                                                                                  #
+            , "dst"     : "stage/lib_Win64" }                                                                                                          #
+        ,   { "command" : "b2.exe"
+            , "args"    : "--with-system --with-date_time --with-random --with-regex link=static runtime-link=shared threading=multi address-model=64" #
+            , "src"     : "stage/lib"                                                                                                                  #
+            , "dst"     : "stage/lib_Win64" }                                                                                                          #
         ]
     },
     
@@ -48,13 +54,13 @@ config = {
     , "source_tag"   : "OpenSSL_1_0_2l"                     # the tag for the branch to create from
     , "build_targets": # a set of targets
         [#targets
-            [ #steps reqired to build certain turget
+            [ #steps reqired to build certain target
                 { "env"     : "C:/Program Files (x86)/Microsoft Visual Studio/2017/Community/VC/Auxiliary/Build/vcvarsall.bat" 
                 , "env_arg" : ["x86"]}, # run batch in "env" with key in "env_arg" in order to get proper evirinment variables
                 { "command" : "git"
                 , "args"    : "clean -fx -d"}, #clean up local git repo
                 { "command" : "perl" 
-                , "args"    : "Configure VC-WIN32 no-asm --prefix=./../openssl_libs/x86"}, # configure the feature in order to build the target
+                , "args"    : "Configure VC-WIN32 no-asm --prefix=./../openssl_libs/Win32"}, # configure the feature in order to build the target
                 { "command" : "cmd.exe" 
                 , "args"    : "/c"
                 , "rel_args": ["ms/do_ms.bat"]}, # another one configuration step
@@ -71,7 +77,7 @@ config = {
             ,   { "command" : "git"
                 , "args"    : "clean -fx -d"}
             ,   { "command" : "perl"
-                , "args"    : "Configure VC-WIN64A --prefix=./../openssl_libs/x64"}
+                , "args"    : "Configure VC-WIN64A --prefix=./../openssl_libs/Win64"}
             ,   { "command" : "cmd.exe"
                 , "args"    : "/c"
                 , "rel_args": ["ms/do_win64a.bat"]} # this step requires ml64.exe
@@ -84,17 +90,52 @@ config = {
     },
     
     "socket.io-client-cpp" :
-    { "active"       : False
+    { "active"       : True
     , "clone"        : False
-    , "build"        : False
+    , "build"        : True
     , "url"          : "https://github.com/socketio/socket.io-client-cpp"
     , "target_branch": "branch_1.6.1"
     , "source_tag"   : "1.6.1"
-    , "build_targets": []},
+    , "build_targets": 
+        [#targets
+            [#steps
+                { "command" : "git"
+                , "args"    : ["clean -fx -d"]}, #clean up local git repo
+                { "command" : "cmake"
+                , "args"    : 
+                    [ "-G \"Visual Studio 15 2017 Win64\""
+                    , "./", "-DBOOST_ROOT=\"./../boost/\""
+                    , "-DBOOST_LIBRARYDIR=\"./../boost/stage/lib_Win64/\""
+                    , "-DOPENSSL_ROOT_DIR=\"./../openssl_libs/Win64/\""
+                    ] 
+                }
+#            ,   { "command" : "msbuild"
+#                , "args"    : [""]}
+            ],
+            [#steps
+                { "command" : "git"
+                , "args"    : ["clean -fx -d"]}, #clean up local git repo
+                { "command" : "cmake"
+                , "args"    :
+                    [ "-G \"Visual Studio 15 2017\""
+                    , "./", "-DBOOST_ROOT=\"./../boost/\""
+                    , "-DBOOST_LIBRARYDIR=\"./../boost/stage/lib_Win32/\""
+                    , "-DOPENSSL_ROOT_DIR=\"./../openssl_libs/Win32/\""
+                    ] 
+                }
+#            ,   { "command" : "msbuild"
+#                , "args"    : [""]}
+            ],
+            [#steps
+                { "command" : "git"
+                , "args"    : ["clean -fx -d"]} #clean up local git repo
+            ]
+        ]
+    },
 
     "depot_tools" :
     { "active"       : False
-    , "clone"        : True
+    , "clone"        : False
     , "build"        : False
     , "url"          : "https://chromium.googlesource.com/chromium/tools/depot_tools.git"
     , "target_branch": None
@@ -126,14 +167,18 @@ def clone_all(target_dir):
             log.info("\"{0}\" skipped - turned off...".format(target))
 
 def build_boost(target_dir):
-    def cleanup(target_dir):
-        for build_target in boost_config["build_targets"]:
-            target_dir_full = target_dir+"/boost/"+build_target["dst"]
-            if os.path.isdir(target_dir_full):
-                shutil.rmtree(target_dir_full)
+    def cleanup(target_dir, all=True):
+        if all is True:
+            for build_target in boost_config["build_targets"]:
+                if "dst" in build_target:
+                    target_dir_full = target_dir+"/boost/"+build_target["dst"]
+                    if os.path.isdir(target_dir_full):
+                        shutil.rmtree(target_dir_full)
+        
         temp_build_dir = target_dir+"/boost/bin.v2"
         if os.path.isdir(temp_build_dir):
             shutil.rmtree(temp_build_dir)
+        
         temp_lib_dir = target_dir+"/boost/stage/lib"
         if os.path.isdir(temp_lib_dir):
             shutil.rmtree(temp_lib_dir)
@@ -147,25 +192,38 @@ def build_boost(target_dir):
     #build b2.exe and bjam.exe at boost folder if necessary
     if not os.path.isfile(target_dir+"/boost/bjam.exe"):
         if not os.path.isfile(target_dir+"/boost/bootstrap.bat"):
-            log.error("booststrap.bat at {0} does not exist".format(target_dir+"\boost"))
+            log.error("booststrap.bat at {0} does not exist".format(target_dir+"/boost"))
             return
         else:
-            git_tools.run(log, target_dir+"/boost/bootstrap.bat", cwd=target_dir+"/boost")
+            result = git_tools.run(log, target_dir+"/boost/bootstrap.bat", cwd=target_dir+"/boost")
     
     #check it is built    
-    if not os.path.isfile(target_dir+"/boost/bjam.exe"):
-        log.error("bjam.exe at {0} does not exist and did not build by booststrap.bat".format(target_dir+"\boost"))
+    if not os.path.isfile(target_dir+"/boost/b2.exe"):
+        log.error("b2.exe at {0} does not exist and did not build by booststrap.bat".format(target_dir+"\boost"))
         return
 
     #run through build targets and process them
     for build_target in boost_config["build_targets"]:
-        target_dir_full = target_dir+"/boost/"+build_target["dst"]
-        source_dir_full = target_dir+"/boost/"+build_target["src"]
+        target_dir_full = target_dir+"/boost/"
+        source_dir_full = target_dir+"/boost/"
+        
+        if "dst" in build_target:
+            target_dir_full += build_target["dst"]
+        else: 
+            target_dir_full = None
+
+        if "src" in build_target:
+            source_dir_full += build_target["src"]
+        else:
+            source_dir_full = None
 
         #the build itself
-        git_tools.run(log, target_dir+"/boost/bjam.exe", build_target["args"], cwd=target_dir+"/boost")
+        result = git_tools.run(log, target_dir+"/boost/"+build_target["command"], build_target["args"], cwd=target_dir+"/boost")
 
         #TODO: check that everything built ok
+
+        if target_dir_full is None or source_dir_full is None:
+            continue
 
         #check presence of the target directory. If not - create
         if not os.path.isdir(target_dir_full):
@@ -179,7 +237,7 @@ def build_boost(target_dir):
         for file in files:
             shutil.move(source_dir_full+"/"+file, target_dir_full+"/"+file)
     
-    cleanup(target_dir)
+    cleanup(target_dir, False)
     return
 
 def build_openssl(target_dir):
@@ -205,11 +263,23 @@ def build_openssl(target_dir):
                     rel_args = step["rel_args"]
                     for arg in rel_args:
                         args += " " + "{0}/{1}".format(openssl_root_dir, arg) 
-                str_result = git_tools.run(log, step["command"], args, cwd=openssl_root_dir, env=_env)
+                result = git_tools.run(log, step["command"], args, cwd=openssl_root_dir, env=_env)
             else:
                 log.error("Uknown step")
 
 def build_socketio(target_dir):
+    socketio_config = config["socket.io-client-cpp"]
+    socketio_root_dir = target_dir + "/socket.io-client-cpp"
+
+    for build_target in socketio_config["build_targets"]:
+        for step in build_target:
+            arguments = ""
+            for arg in step["args"]:
+                if len(arguments) != 0:
+                    arguments+=" "
+                arguments += arg
+            result = git_tools.run(log, step["command"], arguments, cwd=socketio_root_dir)
+        pass
     pass
 
 def build_all(target_dir):
